chore(AIINT-1261) - change number type to integer

+++ src/tools/automations.js
@@ -6,8 +6,8 @@ import { z } from 'zod';
         name: "list_automations",
         description: "List automations in Zendesk",
         schema: {
-          page: z.number().optional().describe("Page number for pagination"),
-          per_page: z.number().optional().describe("Number of automations per page (max 100)")
+          page: z.number().int().optional().describe("Page number for pagination"),
+          per_page: z.number().int().optional().describe("Number of automations per page (max 100)")
         },
         handler: async ({ page, per_page }) => {
           try {
@@ -31,7 +31,7 @@ import { z } from 'zod';
         name: "get_automation",
         description: "Get a specific automation by ID",
         schema: {
-          id: z.number().describe("Automation ID")
+          id: z.number().int().describe("Automation ID")
         },
         handler: async ({ id }) => {
           try {
