+++ src/tools/groups.js
@@ -6,8 +6,8 @@ import { z } from 'zod';
         name: "list_groups",
         description: "List agent groups in Zendesk",
         schema: {
-          page: z.number().optional().describe("Page number for pagination"),
-          per_page: z.number().optional().describe("Number of groups per page (max 100)")
+          page: z.number().int().optional().describe("Page number for pagination"),
+          per_page: z.number().int().optional().describe("Number of groups per page (max 100)")
         },
         handler: async ({ page, per_page }) => {
           try {
@@ -31,7 +31,7 @@ import { z } from 'zod';
         name: "get_group",
         description: "Get a specific group by ID",
         schema: {
-          id: z.number().describe("Group ID")
+          id: z.number().int().describe("Group ID")
         },
         handler: async ({ id }) => {
           try {
