+++ src/tools/search.js
@@ -9,8 +9,8 @@ import { z } from 'zod';
           query: z.string().describe("Search query string"),
           sort_by: z.string().optional().describe("Field to sort by"),
           sort_order: z.enum(["asc", "desc"]).optional().describe("Sort order (asc or desc)"),
-          page: z.number().optional().describe("Page number for pagination"),
-          per_page: z.number().optional().describe("Number of results per page (max 100)")
+          page: z.number().int().optional().describe("Page number for pagination"),
+          per_page: z.number().int().optional().describe("Number of results per page (max 100)")
         },
         handler: async ({ query, sort_by, sort_order, page, per_page }) => {
           try {
