+++ src/tools/triggers.js
@@ -6,8 +6,8 @@ import { z } from 'zod';
         name: "list_triggers",
         description: "List triggers in Zendesk",
         schema: {
-          page: z.number().optional().describe("Page number for pagination"),
-          per_page: z.number().optional().describe("Number of triggers per page (max 100)")
+          page: z.number().int().optional().describe("Page number for pagination"),
+          per_page: z.number().int().optional().describe("Number of triggers per page (max 100)")
         },
         handler: async ({ page, per_page }) => {
           try {
@@ -31,7 +31,7 @@ import { z } from 'zod';
         name: "get_trigger",
         description: "Get a specific trigger by ID",
         schema: {
-          id: z.number().describe("Trigger ID")
+          id: z.number().int().describe("Trigger ID")
         },
         handler: async ({ id }) => {
           try {
